+++ example/scopo_demo/lib/main.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:ansi_escape_codes/ansi_escape_codes.dart' as ansi;
 import 'package:flutter/material.dart';
+import 'package:pkglog/pkglog.dart';
 import 'package:scopo/scopo.dart';
 
 import 'app/app.dart';
@@ -14,37 +15,37 @@ import 'utils/app_environment.dart';
 
 void main() {
   ScopeConfig.isDebug = true;
+  // ScopeConfig.logger.level = LogLevel.all;
 
   //
   // Logging
   //
 
-  final ansiCodesEnabled = !Platform.isIOS;
+  for (final level in LogLevel.values) {
+    final printer = ansi.AnsiPrinter(
+      ansiCodesEnabled: !Platform.isIOS,
+      defaultState: ansi.SgrPlainState(
+        foreground: switch (level) {
+          LogLevel.verbose => const ansi.Color256(ansi.Colors.gray8),
+          LogLevel.debug => const ansi.Color256(ansi.Colors.gray12),
+          LogLevel.info => const ansi.Color256(ansi.Colors.rgb345),
+          LogLevel.warning => const ansi.Color256(ansi.Colors.rgb440),
+          LogLevel.error => const ansi.Color256(ansi.Colors.rgb400),
+          LogLevel.shout => const ansi.Color256(ansi.Colors.rgb550),
+        },
+        background: switch (level) {
+          LogLevel.verbose => null,
+          LogLevel.debug => null,
+          LogLevel.info => null,
+          LogLevel.warning => null,
+          LogLevel.error => null,
+          LogLevel.shout => const ansi.Color256(ansi.Colors.rgb300),
+        },
+      ),
+    );
 
-  ScopeConfig.log.log = _logByPrinter(
-    ansi.AnsiPrinter(
-      ansiCodesEnabled: ansiCodesEnabled,
-      defaultState: const ansi.SgrPlainState(
-        foreground: ansi.Color256(ansi.Colors.gray8),
-      ),
-    ),
-  );
-  ScopeConfig.logInfo.log = _logByPrinter(
-    ansi.AnsiPrinter(
-      ansiCodesEnabled: ansiCodesEnabled,
-      defaultState: const ansi.SgrPlainState(
-        foreground: ansi.Color256(ansi.Colors.gray16),
-      ),
-    ),
-  );
-  ScopeConfig.logError.log = _logByPrinter(
-    ansi.AnsiPrinter(
-      ansiCodesEnabled: ansiCodesEnabled,
-      defaultState: const ansi.SgrPlainState(
-        foreground: ansi.Color256(ansi.Colors.rgb500),
-      ),
-    ),
-  );
+    ScopeConfig.logger[level].print = printer.print;
+  }
 
   //
   // Scope timeouts
@@ -88,17 +89,3 @@ void main() {
     ),
   );
 }
-
-void Function(String?, String?, Object?, StackTrace?) _logByPrinter(
-  ansi.AnsiPrinter printer,
-) =>
-    (source, message, error, stackTrace) {
-      printer.print(
-        ScopeLog.buildDefaultMessage(
-          source,
-          message,
-          error: error,
-          stackTrace: stackTrace,
-        ),
-      );
-    };