+++ lib/scopo.dart
@@ -2,7 +2,7 @@
 /// dependency injection, asynchronous initialization and disposal.
 library;
 
-export 'src/environment/scope_config.dart' hide RawScopeLogCallback, d, e, i;
+export 'src/environment/scope_config.dart' hide log;
 export 'src/scope/scope.dart';
 export 'src/utils/compare_utils.dart';
 export 'src/utils/is_building.dart';